Create d2c_hack_2.py
Mohit has three types of sweets. Type A is not tasty but has an antidote, B is tasty and also has an antidote and type C is tasty and poisonous.

Eating poisonous sweet results in a stomach-ache
Eating another poisonous sweet, while already having a stomach-ache, results to death
Eating a sweet with antidote can cure the stomachache
Mohit wants to live and eat the maximum number of tasty sweets.

Find the maximum number of tasty sweets that he can have.